refactor: tests & shared code
