+++ packages/docusaurus-plugin-content-blog/src/__tests__/__fixtures__/website-blog-with-last-update/blog/author.md
@@ -2,6 +2,8 @@
 title: Author
 slug: author
 author: ozaki
+last_update:
+  author: seb
 ---
 
 author
