+++ packages/docusaurus-plugin-content-blog/src/__tests__/__fixtures__/website-blog-with-last-update/blog/both.md
@@ -2,7 +2,9 @@
 title: Both
 slug: both
 date: 2020-01-01
-last_update: 2021-01-01
+last_update:
+  date: 2021-01-01
+  author: seb
 author: ozaki
 ---
