+++ packages/docusaurus-plugin-content-blog/src/__tests__/__fixtures__/website-blog-with-last-update/blog/lastUpdateDate.md
@@ -2,7 +2,8 @@
 title: Last update date
 slug: lastUpdateDate
 date: 2020-01-01
-last_update: 2021-01-01
+last_update: 
+  date: 2021-01-01
 ---
 
 last update date
