+++ packages/docusaurus-plugin-content-blog/src/__tests__/__fixtures__/website-blog-with-tags/blog/another-simple-slug-with-tags.md
@@ -6,7 +6,6 @@ date: 2020-08-15
 author: Sébastien Lorber
 author_title: Docusaurus maintainer
 author_url: https://sebastienlorber.com
-last_updated: 2022-08-15
 
 tags: [tag1]
 ---
